Keep html.parser and urllib.request off the CLI import
